refactor shared logic into shared functions
